nvme-cli/spdk: add the support for the Intel plugin commands
With this patch, the below Intel plugin commands can be supported
by SPDK. The most useful commands are internal-log and smart-log-add
which can be used to collect some common SSD logs.

Intel vendor specific extensions

The following are all implemented sub-commands:
  id-ctrl         Send NVMe Identify Controller
  smart-log-add   Retrieve Intel SMART Log, show it
  market-name     Retrieve Intel Marketing Name log, show it
  temp-stats      Retrieve Intel Temperature Statistics log, show it
  lat-stats       Retrieve Intel IO Latency Statistics log, show it
  internal-log    Retrieve Intel internal firmware log, save it
  version         Shows the program version
  help            Display this help

+++ nvme-ioctl.c
@@ -91,7 +91,7 @@ int nvme_get_nsid(int fd)
 
 int nvme_submit_passthru(int fd, int ioctl_cmd, struct nvme_passthru_cmd *cmd)
 {
-	if (ioctl_cmd == NVME_IOCTL_ADMIN_CMD) {
+	if (ioctl_cmd == (int)NVME_IOCTL_ADMIN_CMD) {
 		if (nvme_spdk_is_valid_fd(fd) == 0) {
 			return nvme_spdk_submit_cmd_passthru(fd, cmd, true);
 		}
